尝试加入字典

+++ src/main/java/com/agoni/dgy/controller/MajorController.java
@@ -4,6 +4,7 @@ package com.agoni.dgy.controller;
 import com.agoni.dgy.model.bo.MajorSearchFrom;
 import com.agoni.dgy.model.po.Major;
 import com.agoni.dgy.service.MajorService;
+import com.baomidou.mybatisplus.core.conditions.Wrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,7 +12,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
@@ -28,6 +32,7 @@ public class MajorController {
 
     @Autowired
     private MajorService majorService;
+    
     
     @GetMapping
     public ResponseEntity<IPage> majorPage(@Validated MajorSearchFrom from) {
@@ -49,7 +54,7 @@ public class MajorController {
     
     @DeleteMapping
     public ResponseEntity<Boolean>  delete(@RequestBody List<Major> majorList){
-        List<Integer> ids = majorList.stream().map(Major::getId).collect(Collectors.toList());
+        List<Long> ids = majorList.stream().map(Major::getId).collect(Collectors.toList());
         boolean b = majorService.removeByIds(ids);
         return new ResponseEntity<>(b, HttpStatus.OK);
     }
